requirements.txt
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
tqdm>=4.60.0
requests>=2.28.0
PyMuPDF>=1.24.0
pymupdf4llm>=0.3.0
sentence-transformers>=2.2.0
chromadb>=0.5.0
google-genai>=1.0.0
